Inline locals.js into server.js
